long

+++ src/Data.java
@@ -1,12 +1,12 @@
 import java.io.Serializable;
 
 public class Data implements Serializable {
-    int n;
-    int m;
-    int i0;
-    int k;
+    long n;
+    long m;
+    long i0;
+    long k;
 
-    public Data(int n, int m, int i0, int k) {
+    public Data(long n, long m, long i0, long k) {
         this.n = n;
         this.m = m;
         this.i0 = i0;
